refactor(next/swc): rename references
